requirements.txt
basicsr==1.4.2
dlib==19.24.4
einops==0.3.0
facexlib==0.3.0
numpy==1.26.4
omegaconf==2.1.1
open_clip_torch==2.24.0
opencv_contrib_python==4.9.0.80
opencv_python==4.9.0.80
Pillow==9.0.1
Pillow==10.3.0
pytorch_lightning==1.4.2
scipy==1.13.1
timm==0.9.12
torch==1.11.0
torchvision==0.12.0
tqdm==4.65.2
transformers==4.38.0.dev0
xformers==0.0.26.post1
torchaudio==2.2.2+cu118
accelerate==0.28.0
ftfy==6.2.0
regex==2023.12.25
python-dateutil==2.9.0.post0
pytorch-lightning==2.2.1 # only for loading pretrained sd weight
tensorboard==2.16.2 # for tensorboard event visualization
protobuf==4.25.3 # for tensorboard
lpips==0.1.4
gradio==4.25.0